bug fixes, refactor
